Shots reflect on circles and apply damage
+ Shots now reflect off the head when colliding
+ Shtos apply damage to the head and itself when colliding

Note: My vision for this screen saver is to apply just a touch of name learning. For that I envision that when a head is popped, a copy of it will appear on the side along with a note that says "Denise was defeated" for example for Denise. That way people are learning names. Then, when there is one head left, the screen saver stops, the head floats to the middle of the screen, is enlarged, and a note says "Ryan is the winner" (for example). Then the screen saver should wait about 15 seconds and then start over with a new batch of heads.